requirements.txt
decorator==4.1.2
Django==2.0
django-extensions==1.9.7
djangorestframework==3.7.3
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
mypy==0.550
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
pkg-resources==0.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psutil==5.4.1
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
Pygments==2.2.0
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
tox==2.9.1
traitlets==4.3.2
typed-ast==1.1.0
typing==3.6.2
virtualenv==15.1.0
wcwidth==0.1.7
